% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAxisPath.R
\name{CalculateAxisPath}
\alias{CalculateAxisPath}
\title{Calculate Axis Path}
\usage{
CalculateAxisPath(var.names, min, max)
}
\arguments{
\item{var.names}{list of variables to be plotted on radar plot}

\item{min}{MININUM value required for the plotted axes (same value will be applied to all axes)}

\item{max}{MAXIMUM value required for the plotted axes (same value will be applied to all axes)}
}
\value{
a dataframe of the calculated axis paths
}
\description{
Calculates x-y coordinates for a set of radial axes (one per variable being plotted in radar plot)
}
