% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcCircleCoords.R
\name{funcCircleCoords}
\alias{funcCircleCoords}
\title{Generate circle coordinates}
\source{
Adapted from Joran's response to \url{http://stackoverflow.com/questions/6862742/draw-a-circle-with-ggplot2}.
}
\usage{
funcCircleCoords(center = c(0, 0), r = 1, npoints = 100)
}
\arguments{
\item{center}{coordinate for centroid}

\item{r}{radius}

\item{npoints}{number of coordinates to generate}
}
\value{
a dataframe
}
\description{
Generate coordinates to draw a circle.
}
