% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_radar_point.R
\name{geom_radar_point}
\alias{geom_radar_point}
\title{Radar points for polar plot}
\usage{
geom_radar_point(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and \code{inherit.aes = TRUE} (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer.}

\item{position}{A position adjustment to use on the data for this layer.}

\item{...}{Other arguments passed on to \code{layer()}'s params argument.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them.}
}
\description{
For continuous scale ensures that first and last points do not share same angular position. For discrete scale it is equivalent to \code{geom_point}.
}
\examples{
example_df <- data.frame(x = factor(c("One", "Two", "Three")),
                         y = c(0.2, 0.4, 1.0))
ggplot2::ggplot(example_df, ggplot2::aes(x = x, y = y)) +
  geom_radar_point() +
  ggplot2::coord_polar() +
  ggplot2::ylim(0, 1)
}
