% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggradar-package.R, R/ggradar.R
\docType{package}
\name{ggradar-package}
\alias{ggradar2}
\alias{ggradar2}
\title{ggradar: Create radar charts using ggplot2}
\source{
Most of the code is from \url{http://rstudio-pubs-static.s3.amazonaws.com/5795_e6e6411731bb4f1b9cc7eb49499c2082.html}.
}
\usage{
ggradar(
  plot.data,
  base.size = 15,
  font.radar = "sans",
  values.radar = c("0\%", "50\%", "100\%"),
  all.radar=NULL,
  all.label=NULL,
  axis.labels = colnames(plot.data)[-1],
  grid.min = ifelse(length(all.radar),min(all.radar),0),
  grid.mid = 0.5,
  grid.max = ifelse(length(all.radar),max(all.radar),1),
  centre.y = grid.min - ((1/9) * (grid.max - grid.min)),
  plot.extent.x.sf = 1,
  plot.extent.y.sf = 1.2,
  x.centre.range = 0.02 * (grid.max - centre.y),
  label.centre.y = FALSE,
  grid.line.width = 0.5,
  gridline.min.linetype = "soild",
  gridline.mid.linetype = "longdash",
  gridline.max.linetype = "longdash",
  gridline.min.colour = "grey",
  gridline.mid.colour = "#007A87",
  gridline.max.colour = "grey",
  grid.label.size = 6,
  gridline.label.offset = -0.1 * (grid.max - centre.y),
  label.gridline.min = TRUE,
  label.gridline.mid = TRUE,
  label.gridline.max = TRUE,
  axis.label.offset = 1.15,
  axis.label.size = 5,
  axis.line.colour = "grey",
  group.line.width = 1.5,
  group.point.size = 6,
  group.colours = NULL,
  background.circle.colour = "#D7D6D1",
  background.circle.transparency = 0.2,
  plot.legend = if (nrow(plot.data) > 1) TRUE else FALSE,
  legend.title = "",
  plot.title = "",
  legend.text.size = 14,
  legend.position = "left",
  fill = FALSE,
  fill.alpha = 0.5
)
}
\arguments{
\item{plot.data}{dataframe comprising one row per group}

\item{base.size}{text size}

\item{font.radar}{text font family}

\item{values.radar}{values to print at minimum, 'average', and maximum gridlines}

\item{all.radar}{values to print each gridlines}

\item{all.label}{values to print each gridlines labels}

\item{axis.labels}{names of axis labels if other than column names supplied via plot.data}

\item{grid.min}{value at which mininum grid line is plotted}

\item{grid.mid}{value at which 'average' grid line is plotted}

\item{grid.max}{value at which maximum grid line is plotted}

\item{centre.y}{value of y at centre of plot}

\item{plot.extent.x.sf}{controls relative size of plot horizontally}

\item{plot.extent.y.sf}{controls relative size of plot vertically}

\item{x.centre.range}{controls axis label alignment}

\item{label.centre.y}{whether value of y at centre of plot should be labelled}

\item{grid.line.width}{width of gridline}

\item{gridline.min.linetype}{line type of minimum gridline or all gridline}

\item{gridline.mid.linetype}{line type of 'average' gridline}

\item{gridline.max.linetype}{line type of maximum gridline}

\item{gridline.min.colour}{colour of minimum gridline or all gridline}

\item{gridline.mid.colour}{colour of 'average' gridline}

\item{gridline.max.colour}{colour of maximum gridline}

\item{grid.label.size}{text size of gridline label}

\item{gridline.label.offset}{displacement to left/right of central vertical axis}

\item{label.gridline.min}{whether or not to label the mininum gridline}

\item{label.gridline.mid}{whether or not to label the 'mininum'average' gridline}

\item{label.gridline.max}{whether or not to label the maximum gridline}

\item{axis.label.offset}{vertical displacement of axis labels from maximum grid line, measured relative to circle diameter}

\item{axis.label.size}{text size of axis label}

\item{axis.line.colour}{colour of axis line}

\item{group.line.width}{line width of group}

\item{group.point.size}{point size of group}

\item{group.colours}{colour of group}

\item{background.circle.colour}{colour of background circle/radar}

\item{background.circle.transparency}{transparency of background circle/radar}

\item{plot.legend}{whether to include a plot legend}

\item{legend.title}{title of legend}

\item{plot.title}{title of radar plot}

\item{legend.text.size}{text size in legend}

\item{legend.position}{position of legend, valid values are "top", "right", "bottom", "left"}

\item{fill}{whether to fill polygons}

\item{fill.alpha}{if filling polygons, transparency values}
}
\value{
a ggplot object
}
\description{
This package exports the functionality of ggplot2 to create radar charts.
}
\examples{
\dontrun{
library(ggradar)
library(dplyr)
library(scales)
library(tibble)

set.seed(21)
sampleA <- sample(seq(1,23),size=23)
sampleB <- sample(seq(1,23),size=23)
data <- data.frame(group=c('sampleA','sampleB'),
                   t(data.frame(sampleA=sampleA,sampleB=sampleB)))
chroms <- paste0('chr',c(seq(1:22),'X'))
colnames(data)[-1] <- chroms
data
ggradar2(data,all.radar=seq(1:23),axis.label.size=3.5,grid.label.size=4,group.point.size=1,
         group.line.width=0.6, legend.position='right')
}
}
\author{
\strong{Maintainer}: Ricardo Bion \email{ricardo.bion@airbnb.com}

}
\keyword{internal}
